refactor: ReservationTime 패키지 수정
