+++ src/main/java/roomescape/domain/reservationtime/admin/dto/CreateTimeRequest.java
@@ -1,4 +1,4 @@
-package roomescape.domain.reservationtime.admin;
+package roomescape.domain.reservationtime.admin.dto;
 
 import jakarta.validation.constraints.NotNull;
 import java.time.LocalTime;
