+++ src/main/java/roomescape/domain/reservationtime/admin/dto/CreateTimeResponse.java
@@ -1,17 +1,17 @@
-package roomescape.domain.reservationtime.admin;
+package roomescape.domain.reservationtime.admin.dto;
 
 import com.fasterxml.jackson.annotation.JsonFormat;
 import java.time.LocalTime;
 import roomescape.domain.reservationtime.ReservationTime;
 
-public record ReservationTimeResponse(
+public record CreateTimeResponse(
     Long id,
     @JsonFormat(pattern = "HH:mm")
     LocalTime startAt
 ) {
 
-    public static ReservationTimeResponse from(ReservationTime reservationTime) {
-        return new ReservationTimeResponse(
+    public static CreateTimeResponse from(ReservationTime reservationTime) {
+        return new CreateTimeResponse(
             reservationTime.getId(),
             reservationTime.getStartAt()
         );
