+++ src/main/java/roomescape/domain/reservationtime/admin/dto/ReservationTimeResponse.java
@@ -1,17 +1,17 @@
-package roomescape.domain.reservationtime.admin;
+package roomescape.domain.reservationtime.admin.dto;
 
 import com.fasterxml.jackson.annotation.JsonFormat;
 import java.time.LocalTime;
 import roomescape.domain.reservationtime.ReservationTime;
 
-public record CreateTimeResponse(
+public record ReservationTimeResponse(
     Long id,
     @JsonFormat(pattern = "HH:mm")
     LocalTime startAt
 ) {
 
-    public static CreateTimeResponse from(ReservationTime reservationTime) {
-        return new CreateTimeResponse(
+    public static ReservationTimeResponse from(ReservationTime reservationTime) {
+        return new ReservationTimeResponse(
             reservationTime.getId(),
             reservationTime.getStartAt()
         );
