+++ src/test/java/roomescape/domain/reservationtime/admin/AdminReservationTimeControllerTest.java
@@ -24,6 +24,9 @@ import org.springframework.http.MediaType;
 import org.springframework.test.context.bean.override.mockito.MockitoBean;
 import org.springframework.test.web.servlet.MockMvc;
 import roomescape.domain.reservationtime.ReservationTimeService;
+import roomescape.domain.reservationtime.admin.dto.CreateTimeRequest;
+import roomescape.domain.reservationtime.admin.dto.CreateTimeResponse;
+import roomescape.domain.reservationtime.admin.dto.ReservationTimeResponse;
 import roomescape.support.auth.AdminRequestValidator;
 import roomescape.support.exception.ConflictException;
 import roomescape.support.exception.errors.ReservationTimeErrors;
